// src/types/enums.ts
// 性别枚举
export enum SexEnum {
    Male = '男',
    Female = '女',
    MaleCode = 1,
    FemaleCode = 2,
}

// 行政区划级别枚举
export enum AreaLevelEnum { Province = 1, City = 2, County = 3 }

// 运营商枚举
export enum TELCO_ENUM {
    CT = "电信",
    CM = "移动",
    CU = "联通",
    CRT = "广电",
}